+++ packages/renderer/tests/components.test.jsx
@@ -8,7 +8,7 @@ import {
   Document,
   Page,
   Text,
-} from '@react-pdf/renderer';
+} from '@nutshelllabs/renderer';
 
 const TestDocument = ({ title = 'Default' }) => (
   <Document title={title}>