+++ packages/renderer/tests/pageWrap.test.jsx
@@ -7,7 +7,7 @@ import {
   Text,
   Image,
   StyleSheet,
-} from '@react-pdf/renderer';
+} from '@nutshelllabs/renderer';
 import renderToImage from './renderComponent';
 
 const styles = StyleSheet.create({