+++ packages/renderer/tests/resume.test.jsx
@@ -9,7 +9,7 @@ import {
   Image,
   Document,
   StyleSheet,
-} from '@react-pdf/renderer';
+} from '@nutshelllabs/renderer';
 import renderToImage from './renderComponent';
 
 const headerStyles = StyleSheet.create({