+++ packages/renderer/tests/svg.test.jsx
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
 
-import { Document, Page, Svg, Font, Text, Tspan } from '@react-pdf/renderer';
+import { Document, Page, Svg, Font, Text, Tspan } from '@nutshelllabs/renderer';
 import renderToImage from './renderComponent';
 import { Tiger, Chart, Chart2, Gradients, Dasharrays } from './svgs';