+++ packages/layout/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nutshelllabs/layout
 
+## 4.4.3
+
+### Patch Changes
+
+- [`d7e7b9d5`](https://github.com/diegomura/react-pdf/commit/d7e7b9d56de69310db3b1949de2637e2ae204582) - fix(layout): filter out empty string fragments in getAttributedString
+
+- Updated dependencies [[`16cf5bd7`](https://github.com/diegomura/react-pdf/commit/16cf5bd7006ee1f64d37d3b9ee72e576ee03e343), [`5b6a6a4a`](https://github.com/diegomura/react-pdf/commit/5b6a6a4a67fc5156b5d937e4a59a4ab9e5150785), [`2ca8ae83`](https://github.com/diegomura/react-pdf/commit/2ca8ae83fcc113b44b28b5397099bfda9d142212)]:
+  - @nutshelllabs/textkit@6.1.1
+  - @nutshelllabs/fns@3.1.3
+  - @nutshelllabs/stylesheet@6.1.3
+  - @nutshelllabs/types@2.9.3
+
 ## 4.4.2
 
 ### Patch Changes